Added backend files
Backend files are now here. Make sure to run pip and install pymongo and flask, as these are needed to run. Server starts on localhost:5000
MongoDB is feature complete minus a few small changes we may make.
Cassandras lib is indev and will be put up here once i can get it ruining
SQL code was crashing the server once flask was updated, so it needs a rewrite. i can upload it
Convert Queries is almost done, lib just needs to be finished. It may be better to change the page a bit though, more in the pull request
Switch to may need to be removed if its not done in time
Neo4j is just starting